refactor: move defining function of system classes and system vclasses to new files
